load annotations lazily
When the ball next to the body text is clicked, asynchronously fetch
the annotation text and stuff it into the textarea.  While at it,
store the color of the ball (currently, plain, pos and neg are
supported) in the annotation state, and use it to paint the ball
dynamically.

There is no "like/hate" control to change the color of the ball yet (a
ball to show a filled annotation is shown in "pos" color, though).