Fix missing "trunk_ref" in duplicate implementation
After registering a cloned document as a new child in the parent
document of the original, we forgot to .put() the new version of the
parent document after updating the trunk_ref property in it.  This
resulted in sporadic error that tried to dereference None.

Also update the visit state before switching into the edit mode
to edit the cloned document; the visit stack should end at the
updated parent that points at this new clone for the redirect to
view_doc that happens after the edit form is submit to work
correctly.