TrunkModel.setHead(): allow caching the information in Trunk objects
To work around inflexibility of the query language of the AppEngine
datastore, some attributes that are useful in queries from the tip
document need to be cached in the trunk objects.

This introduces setHead() meathod to TrunkModel so that we can grab
fields from DocModel whenever trunk.head is updated.

We also need to call setHead() from codepaths that update fields in
the DocModel thar are at the tip of the trunk.