Pageable and Searchable Link Picker
This introduces a pageable and searchable link picker.  It

 - keeps the "New Document" feature introduced earlier;

 - limits the output to show only a handful of links (currently
   hardcoded to 8 on the JavaScript side) at a time;

 - allows the links to be paged with Prev/Next;

 - adds "Search" feature to limit the list to documents whose
   title _BEGINS_ with the given string.

Using head match, not substring match, is a very deliberate choice,
primarily coming from the AppEngine datastore limitation.

The query to enumerate the target of DocLink was rewritten to iterate
over TrunkModel; this change conflicts with what is done in i64
branch (r244), and supersedes it.