Add subscription-related models and task handler
This adds models to support subscriptions to changes per user,
a hook to record changes to be notified, another hook to automatically
subscribe an editor to the modified page as a subscriber, and a
task to scan the changes and summarize them to be sent out as an e-mail
to the subscriber.

There is no UI nor detailed summary yet, but we have to start somewhere.