notification: semantic fixes after review
This fixes the implementation of ChangesSeen class that is used to
keep track of, for each subscribed <user, trunk> pair, the latest
timestamp that records the last time the changes to the page has been
notified.

Earlier, the first change to a trunk after a user subscribed to it
caused all the changes ever made to the trunk to be sent.  We now
create a new ChangesSeen record for the <user, trunk> that points at
the document at the tip when a new subscription is made, so that we
will notify the user only about new changes made to the trunk after
that revision.

ChangesSeen used to incorrectly declare timestamp with auto_now, which
caused copying the timestamp from the last SubscriptionNotification
record not to work correctly (it was overwritten with the current time).

The code in notify.notifyUser() that notifies a single user about
recent changes assumed that the invariant for models.ChangesSeen to
have at most one row per <user, trunk> tuple may be broken, but didn't
properly protect itself from such an invariant violation.