Notify: setSubscription
We already had auto_subscribe in the library module, which was only
used to subscribe the user to the current page when it was edited.
Move the bulk of the logic to a new method in notify module, so that
we can use it also to unsubscribe the user from the page.